requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiomysql
bcrypt
python-jose[cryptography]
python-multipart
streamlit
langchain
langchain-openai
langgraph
python-dotenv
pydantic
